rearrange menus and providers, text-indent customizable
